+++ stories/simpleEditorStories.js
@@ -8,4 +8,12 @@ storiesOf('SimpleEditor', module)
 	<div style={editorWrapperStyle}>
 		<SimpleEditor placeholder="Begin writing..." />
 	</div>
+))
+.add('withContent', () => (
+	<div style={editorWrapperStyle}>
+		<SimpleEditor
+			placeholder="Begin writing..."
+			initialHtmlString="<p><strong>hello this is </strong></p><p><strong>how I type</strong></p>"
+		/>
+	</div>
 ));